requirements.txt
torch==2.0.1
torchvision==0.15.1
torchaudio==2.0.1
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74
Pillow==10.0.0
albumentations==1.3.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
tensorboard==2.13.0
wandb==0.15.8
ultralytics==8.0.200
fastapi==0.103.0
uvicorn==0.23.2
pytest==7.4.0
pytest-cov==4.1.0
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

